requirements.txt
Flask==3.1.1
nltk==3.9.1
pdfminer.six==20250506
python-docx==1.1.2
python-pptx==1.0.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
transformers==4.52.4
Werkzeug==3.1.3
sentencepiece==0.2.0
safetensors==0.5.3
protobuf==6.31.1
filelock==3.18.0
requests==2.32.3
tqdm==4.67.1
numpy==2.3.0
blobfile==3.0.0